Refactor: stats.py を新設し ReindexStats / VaultStats を集約
schemas.py から集計状態・スキーマメタ型を stats.py へ移動。
observable change ゼロ (MCP wire 仕様・既存テスト全通過)。
Agent DX cluster (#38/#20/#80/#39) の後続 PR が schemas.py を
touch せずに stats.py を拡張できる基盤。
