+++ src/vault_search/schemas.py
@@ -146,31 +146,3 @@ class FolderCount(BaseModel):
         ),
     )
     count: int = Field(description="このフォルダ直下のノート数")
-
-
-# ---------------------------------------------------------------------------
-# Maintenance
-# ---------------------------------------------------------------------------
-
-
-class ReindexStats(BaseModel):
-    """`vault_reindex` ツールのレスポンス: 処理件数の内訳."""
-
-    model_config = ConfigDict(extra="forbid")
-
-    added: int = Field(description="新規追加されたノート数")
-    updated: int = Field(description="更新されたノート数 (mtime 進行)")
-    deleted: int = Field(description="削除されたノート数 (ファイル消失)")
-    skipped: int = Field(description="mtime 変化なしでスキップされたノート数")
-    errors: int = Field(description="パース失敗したノート数")
-
-
-class VaultStats(BaseModel):
-    """`vault_stats` ツールのレスポンス: インデックス全体の統計."""
-
-    model_config = ConfigDict(extra="forbid")
-
-    total_notes: int = Field(description="インデックス済みノート総数")
-    db_size_bytes: int = Field(description="SQLite DB ファイルのサイズ (バイト)")
-    db_size_mb: float = Field(description="SQLite DB ファイルのサイズ (MB, 小数2桁)")
-    vault_root: str = Field(description="Vault ルートの絶対パス")
