+++ src/vault_search/stats.py
@@ -1,0 +1,32 @@
+"""インデックスの集計状態 / スキーマメタを表す型を集約する.
+
+検索結果や個別ノートを表す型は schemas.py に残る。
+将来的に境界が揺らぐ可能性あり (PR 1b/1c で再評価)。
+"""
+
+from __future__ import annotations
+
+from pydantic import BaseModel, ConfigDict, Field
+
+
+class ReindexStats(BaseModel):
+    """`vault_reindex` ツールのレスポンス: 処理件数の内訳."""
+
+    model_config = ConfigDict(extra="forbid")
+
+    added: int = Field(description="新規追加されたノート数")
+    updated: int = Field(description="更新されたノート数 (mtime 進行)")
+    deleted: int = Field(description="削除されたノート数 (ファイル消失)")
+    skipped: int = Field(description="mtime 変化なしでスキップされたノート数")
+    errors: int = Field(description="パース失敗したノート数")
+
+
+class VaultStats(BaseModel):
+    """`vault_stats` ツールのレスポンス: インデックス全体の統計."""
+
+    model_config = ConfigDict(extra="forbid")
+
+    total_notes: int = Field(description="インデックス済みノート総数")
+    db_size_bytes: int = Field(description="SQLite DB ファイルのサイズ (バイト)")
+    db_size_mb: float = Field(description="SQLite DB ファイルのサイズ (MB, 小数2桁)")
+    vault_root: str = Field(description="Vault ルートの絶対パス")
